Add meta-jobs.
If a job is named with a regex starting with ^, treat it as a
meta-job and apply its attributes to matching jobs.

Explicitly set attributes on normal job definitions will override
those set by a meta job.

This should reduce the need to copy/paste error messages for
common jobs.

Change-Id: I4362f4892d1c60514fbaa9092ecd8b44493c7421