Create patch set 4
Patch Set 4:

In testing it, I realized there's a subtlety around the job_urls variable -- it only gets filled in when the job is complete.  I went ahead and added a bit to try to find the current build record if one is running and use the URL for that.  (The build records are ephemeral, so it's not always there.)

Also, I rejiggered the HTML so the name of the job is the link.

Whadya think?

Patch-set: 4
Subject: Add links to jobs in zuul status page.
Commit: 6987883f69db4e90fdb0a97a61646e50f524b81a
Label: Code-Review=0