Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

Cool.  But that method is used by both the log and the status page, which is HTML.  I think it should be given an "html=False" default argument, and then called with html=True when it's being used in the status page.

When it's false, use the current output, when it's html, add the link in (with a real html anchor).  Currently the whole status page is in a <pre>; we probably want to get rid of that and do a little more html formatting.

Patch-set: 2
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0