Update patch set 3
Patch Set 3:

The problem with only checking this in possiblyReportChange is
that method is only called when all the jobs for a change are
complete.  That makes this only a minor improvement on the
current case.  Currently jobs are only canceled when the change
at the head of the queue is complete.  This patch will
additionally cancel jobs behind any change in the queue, once
that change is complete.  That's better, but we can actually
cancel jobs for changes behind any change once any individual
build for that change fails.  For that reason, I think we need to
invoke the process of looking for that situation from
onBuildCompleted.

Also, within didAnyJobFail, I'm leaning toward having the check
be "and build.result == 'FAILURE'".  For what we know we're going
to use this method for now, SKIPPED implies at least one other
job FAILED (the gate job for us), and CANCELED implies this has
already been done.  In other words, I think just literally
testing for failure will work for us, and of course, is closer to
what the method name describes.

Patch-set: 3
Label: Code-Review=-1
Label: Workflow=0