Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

As we discussed on IRC, if it's all the same to you, I'd rather just wait for mtaylor to finish PBR and use that instead.  It will mean much less additional code in the zuul repo.

Patch-set: 3
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0