Clear the git repo cache on update.
The GitPython Repo object caches whether a given sha is held in an packed
or loose database.  If the remote repository is repacked, that information
could be invalid.  Therefore, each time the local repo is updated from the
remote, clear the object database cache.

See this file for the cache in action:
  https://github.com/gitpython-developers/gitdb/blob/master/gitdb/db/base.py#L227

Change-Id: I547ba01d8cdc20f431e93e02736b498f744992c0