fix back compat issues with python modules
When using python-daemon 1.6, the interface has change in an
uncompatible way. Clark pointed me to Gerritbot which solve that issue
with a simple try / catch block implemented with:

https://github.com/openstack-ci/gerritbot/commit/b2be72e69d08c774989b2cc8c39748ed06ffb1bf

So this patch is merely a copy/paste from David "davido" Ostrovsky with
a small workaround for pyflake issue #13 (we have to prented we are
using the variable holding the module).

Change-Id: Iffdf7fca067734fa9c09b5bddfb13f122e6251a7