Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

This looks good to me; however, I removed the --ignore line and the errors reported by E125 do not appear to be false alarms to me. It is saying that when a conditional for a block is continued over multiple lines that continuation should not be at the same indentation level as the block code. It should have an extra level of indentation. See http://paste.openstack.org/show/21067/ for an example.

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0