raise exception when info reference cant be found
We always assume being able to get the references from a specific URL
although that might not be available (for example if Apache is not
properly setup). Logging an error with the URL we attempted to reach and
raising the urllib2 exception will give an obvious hint to the
administrator.

Change-Id: I0465f8ec0749f6f16d64ddb22dd2ad93dbc42c82