Update patch set 1
Patch Set 1:

This looks great.  In the long run, we want to consume gerritlib directly, in case you have interest in doing more work in this area.

We've frozen zuul around the openstack release, I can merge this to master next week, or to the "dev" branch now (if you want to re-propose it there).

Patch-set: 1
Label: Verified=0
Label: Workflow=+1