Update patch set 4
Patch Set 4: Looks good to me (core reviewer)

Looks good. I think that the unittests are failing due to a race condition in the execution of mocked tests and the checking of those test results. (this race condition is not a fault of your patch it has been around for a while).

Patch-set: 4
Label: Code-Review=+2
Label: Workflow=0