doc about a post-merge pipeline
Reacting on `change-merged` event definitely requires --force-message to
be passed to `gerrit review` or Gerrit will complain the change is
closed whenever it is also passed --verify X or --code-review Y.

Added a simple example to setup a post-merge independant pipeline.

Left a note about ``change-merged`` lacking the commit sha1. One should
instead use the ``ref-updated`` when the commit sha1 is needed.

Close #1067083

Change-Id: Ic402a0048651d10ac619010fdfd58725374368cd