Retry git fetches to work around GitPython bug.
GitPython when fetching refs may improperly read in git progress info
and interpret it as fetch info. When this happens an AssertionError is
thrown. However, since fetching seems to cache the ref locally
subsequent fetches work without running into the progress info problem.
So fetch again if an AssertionError is caught.

Upstream bugfix
https://github.com/gitpython-developers/GitPython/pull/42

Change-Id: I1768b836aade2853d4dd3b2f4cd1fce7ec035524