let us specify Gerrit root URL
Our code was expecting the Gerrit web interface to be available at the
root of the Gerrit server which is not always the case since people
could publish the Gerrit web interface under a subdirectory. In such
configuration, the existing assumption prevents Zuul from getting
references information.

This patch adds a new optional configuration variable 'baseurl' to the
'gerrit' section. It lets one override the default:

 [gerrit]
 server=gerrit.wikimedia.org
 baseurl=https://gerrit.wikimedia.org/r/

Updated Gerrit._getInfoRefs() to use the new parameter, still failing
back to the old default whenever gerrit.baseurl is not set.

Change-Id: Ia9910acf46fe82c425911784412356598b3c7eac