Update patch set 1
Patch Set 1:

After discussion on IRC, I do not think that we should use the API autoindexing stuff for zuul.  There's no usable documentation in there currently, and even so, it would be developer documentation which we should be careful not to mix with user docs.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0