Update patch set 1
Patch Set 1:

I have tested the change on a dev box and I am no more receiving the errors. Zuul happily triggers jobs for my repository set with a 'production' branch as HEAD.

The change is being deployed in production by puppet as we speak, will check it out in a few.

Patch-set: 1