Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

Can you explain how forge committer affects this change?  I believe the email address that's matched against is the address of the _gerrit account_ that is performing the action.  In the case of a commit, it should be the address of the uploader, not the committer.

Patch-set: 1
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0