Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

I think this is going to run too often.  Because of this line:

  if hasattr(change, 'refspec') and not ref:

It's going to run everytime launchJobs is called when either of the following is true:

 * the event is a ref-updated event
 * the event is a change, and zuul has already made a zuul ref with it merged.

Which is _most_ events, and since launchJobs is idempotent, it gets called a lot more often that it needs to.

In fact, you can even look at the Jenkins unit test run for this change, and you'll see that it took 2:20 while the run for the previous change took 1:20.

I think instead of doing the update here, we should do in a method that's called from process_event_queue, before the event is enqueued into any pipeline.

basically:

  if event is for a ref-update:
    update repo
  ... continue into pipeline matching ...

Patch-set: 1
Label: Verified=0
Label: Code-Review=-1