Update patch set 1
Patch Set 1:

If this is something that we do want to add to zuul I will need to add a proper Event type so that events can be tagged appropriately (this make subscribing to a subset of events easy). And error handling will need to be added to the event proxy thread. Also, documentation will need to be written.

As is this change is the result of me scratching an itch and I am looking for some general approval before I make it mergeable.

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0