Update patch set 5
Patch Set 5:

The debug message could show up the project name. I am not sure how to get it though :(

Another issue I found is that we end up doing a remote update even if none of the pipelines are interested in the change.  I guess we want to update if and only if the change has been enqueued in at least one pipeline.  That will save a few updates on the remote.

Patch-set: 5