Update patch set 5
Patch Set 5: I would prefer that you didn't merge this

This version only sets the zuul refs if it is configured to _push_ the zuul refs to gerrit, but those are two different things.  Setting the zuul refs in the zuul repo always needs to happen, regardless if whether they are then pushed to gerrit.

Patch-set: 5
Label: Verified=0
Label: Code-Review=-1
Label: Workflow=0