Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(13 inline comments)

This is great, exactly how it should be done!

I left some inline comments about small things.

I think we need the following additions to avoid regressing on validation and testing:

1) validate the pipeline names in the project template
2) validate that each parameter required by a template is supplied
3) Add a case(s) for layout validation
4) Add a test case that triggers a job defined by a template

Currently the pipeline names for each project are actually checked in layoutvalidator (so you can't assign a job to a pipeline that doesn't exist), so we want to make sure you can't use a template to assign a job to a pipeline that doesn't exist.  Addressing #1 should be pretty similar to how it's already done for projects.

#2 will probably require new logic.  You could either do it in layout validator, or in the job configuration inside the scheduler, if that's easier.

Patch-set: 1
Label: Verified=0
Label: Code-Review=-1