Gerrit 2.5/2.6 stream-events has new event types
We run Gerrit 2.5 with most of 2.6 cherry picked in. At some point
Gerrit has added some new events to the `stream-events` command:

 - draft-published
 - merge-failed
 - reviewer-added

Each event comes with a field holding the account of the user performing
the action. Since Gerrit uses different names for that field, we have to
maintain a map to find out which key holds the account information.

This patch update the map for the three new events.

Also fixed a minor typo in an error message.

Change-Id: I79032280884e50edf83a3ddf60af7a2f62bde971