Update patch set 11
Patch Set 11:

The gist of my changes in the latest patchset are:

a) Remove setZuulRef because we can just call createZuulRef with the right target when we want it created.

b) Cause getRepo to recreate the python-git repo object before returning, so we don't have to do that explicitly everywhere.

c) Drop the loop at the start of mergeChanges that originally reset the repo, and in the previous patchsets, recreated the repo object and created zuul refs.  Changes a and b obviate that now.

d) Add a comment to remind us why we have to reset the repo before calling getBranchHead.

e) Fix the boolean logic in "c.project != change.project or c.branch != change.branch".  Oops.  The old way was causing it to never be called.

Patch-set: 11
Label: Verified=0