Update patch set 1
Patch Set 1:

No idea if this is the right place to be documenting it, but didn't want to forget to put it somewhere so I submitted this while I was still thinking about it. Open to suggestions if we want this in one of the openstack-infra/config docs instead, or on the wiki.

Patch-set: 1
Reviewer: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0