Update patch set 2
Patch Set 2:

I don't think I want to host a duplicate copy of the code in this repo.  I think the best thing would be to link to a hosted copy as you do for jquery.  Perhaps:

https://raw.github.com/mathiasbynens/jquery-visibility/master/jquery-visibility.min.js

Or secondarily, change both of those back to relative URLs and document the external dependencies needed for the status page.

Patch-set: 2