Update patch set 2
Patch Set 2:

If you really object to statsd, we can look into doing this.  But we'll have to figure out how to express that we want statsd on the openstack servers -- and we want it installed by zuul's setup.py;  In other words, I don't think we've ever tried telling puppet to pip install something with an extra, so we need to figure that out.

I'd like Monty to weigh in on this too.

Having said all that, I actually think that statsd is so simple it should just be a hard requirement.  If you don't configure it, it stays out of your way.  We're not talking about mysql or postgres here.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0