Update patch set 11
Patch Set 11: I would prefer that you didn't merge this

This looks excellent, but I just thought of one more thing that it would be good to make sure we get right the first time (to avoid backwards incompatible changes, etc).

Could we remove the "template:" dict key in "project-templates:" so that it more closely matches the way a "project:" is defined?

So the docs would look like this: http://paste.openstack.org/show/36512/

Thanks so much, and sorry I didn't think of this earlier.

Patch-set: 11
Label: Code-Review=-1
Label: Workflow=0