Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

Is this something that we can't do with a simple grep? Or are we trying to do something more like ensure the yaml is parseable and results in a list of projects that we expect?

This is a simple thing to add, but is it necessary? I don't think this is a hard sell as the feature appears to be simple, but some explanation of the use case would be nice.

Patch-set: 1
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0