Test statsd.
Add a fake statsd listener, and actually listen on a random udp
port for statsd messages.  Configure STATSD env vars to use it
before each test.

Add a test that exercises all of the statsd methods that we use
explicitly.

Add checks of the statsd output to the basic "run jobs" test case
to make sure that we're covering some of the statsd code in the
scheduler.

Change-Id: I3361c4e334155dda413bf343e6f1c6f4e277a3ff