Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

This is my fault for not designing this better to start with, but at this point, large parts of zuul are actually depending on the Change objects in separate pipelines being different objects.  In particular, the Change object in zuul holds information about the current builds, etc, so you can see that if two different pipelines shared a single Change, it would get confusing and you'd have undefined behavior (like submitting a change based on successful jobs in another pipeline when the jobs in this one failed).

At any rate, we definitely _should_ be smarter about querying gerrit, but it's a little more complicated than this, sadly.

(In my defense, when I wrote this, we didn't query gerrit at all, so it didn't matter that we had multiple change objects.)

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0