Update patch set 5
Patch Set 5:

I don't want to break other users lightly.  That's why this is first going into a feature branch, so that we can still land changes to master while we work on this.

But I'd really like to get rid of the current launcher.  It's a mess, and it's unreliable.  It is easily broken by Jenkins REST API changes, it has to make way too many http requests to do simple things, it's fragile and subject to network errors, and the worst thing is that it's hard to set up.  It doesn't bother us that much right now since our systems are running, but a new user to Zuul has to configure _a lot_ of parameters for every job they want to trigger, install the notification plugin, and configure _that_ for every job they want to configure as well.

With this, a new user can just install the gearman plugin for jenkins, configure it once (globally), and all the jobs will be registered with gearman and zuul can invoke them.

Moving our systems will be a little more difficult, but not too bad.  We should only need to install the gearman plugin, install gearman, and once that is done, just upgrade zuul to switch.  Then we can remove the un-needed zuul parameters and notification configuration from jobs.  

Gearman is fairly easy to install, but if would make things simpler, we could even include a simple Gearman server in Zuul itself so that there is not another configuration step.  That could be optional so that larger systems could use dedicated Gearman servers.  Also, using gearman means that you can have multiple Jenkins masters, which is a great new feature, along with easily plugging in new systems to handle tasks that aren't Jenkins.

I'm happy to talk more about this if you want -- we definitely don't want to leave anyone, especially you, out in the cold with this change.  We should probably chat in IRC or email though.  I believe this could be reworked to support both the gearman and jenkins launchers, but I want to convince you that it would be better to drop the jenkins launcher because it will be a better user experience.

Patch-set: 5
Label: Code-Review=0