Update patch set 8
Patch Set 8:

The copied change intended for dup2 stick in the pipeline change queue as confirmed by self.assertEmptyQueues():

 pipeline dup2 queue org/project contents [<Change 0x1069accd0 1,None>]

When dup2 pipeline looks for jobs to run (_findJobsToRun) it will find that the requested job is already being build since dup1 triggered it. That is what the lame print detects:

 FOUND A BUILD <Pipeline dup2> for <Change 0x1069accd0 1,None>

That could be a nice feature to have, similar changes triggering the same job might want to share the build result and thus avoid running the exact same job :D


The copies should only share their project but not the current_build_set. As you said, to be safe, Changeish and Change should implement a copy method that willset project to be a reference and deep copy the rest.

OR, we can take advantage of it and avoid triggering duplicate jobs :-] Some logic could be added onBuildComplete to update all the changes impacted and unqueue them.

Patch-set: 8
Reviewer: Gerrit User 2475 <2475@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0