Update patch set 3
Patch Set 3:

If everything "just works" then this is less effort to manage. A special signal handler in zuul to restart and reattach gearmand without needing to restart zuul itself might be warranted though? That would allow us to trigger independent gearmand restarts from the zuul initscript if that's something we expect would be a regular occurrence.

Patch-set: 3
Label: Workflow=0