Update patch set 3
Patch Set 3:

I like the idea of an independent service. Gearman isn't so tightly coupled to zuul that it cannot run on its own. Though in this case Zuul is coupling itself to gearman so it may make sense here to just join them at the hip.

Patch-set: 3
Label: Workflow=0