Wait for internal gearman server to settle.
If we're using the internal gearman server, wait for workers
to register their functions before proceeding with startup.

This is mostly to handle restarts where we may be immediately
ready to run jobs, but workers may not have registered functions
yet.  In that case, we would immediately start declaring jobs
LOST.  So give them a chance to show up first.

The gearman-plugin for jenkins currently tries to reconnect
every 3 seconds.  We could probably even lower that to 1 or 2.

Change-Id: I173b16ba78cecac91acec26b4f2d55f38610e5a2