Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

Would it make sense to resuscitate a call to the Jenkins API to fetch the jobs ?

A call to /api/json?tree=jobs%5Bname,buildable%5D gives out the list of jobs and does not seem to trigger lazy build initialization in Jenkins.  But that might not work with multiple master configuration.

Example:

 https://integration.wikimedia.org/ci/api/json?tree=jobs%5Bname,buildable%5D&pretty

Buildable being set to false when the job is disabled.

Patch-set: 3