Update patch set 3
Patch Set 3:

Zuul itself can (and does) ask Gearman what jobs are registered, so this isn't actually solving a runtime problem.  Instead, this is so that we can automatically validate changes to our layout.yaml file to make sure they reference jobs that are defined in our JJB config (and do all of that before the changes are merged and the jobs exist).

Patch-set: 3
Label: Code-Review=0