Update patch set 4
Patch Set 4:

After thinking about this a bit, I'm not sure we should support updating the build name and numbers; they are used to asynchronously set build descriptions or stop running builds, which means that if they are updated, there is a race condition where zuul may try to stop a build using the old name or number.  Perhaps it's better to declare that they are immutable, and just let the url change?

Patch-set: 4
Label: Code-Review=0
Label: Workflow=0