Fetch specific refs on ref-updated events
The current behavior is that for every event, run
'git remote origin update', which is quite a bit of overhead and
doesn't match what the comments say should be happening.  The goal
is to ensure that when new tags arrive, we have them locally in
our repo.  It's also not a bad idea for us to keep up with remote
branch movements as well.

This updates the event pre-processor to fetch the ref for each
ref-updated event as they are processed.  This is much faster than
the git remote update that was happening before.  It also adds
a git remote update to the repo initialization step so that when
Zuul starts, it will pick up any remote changes since it last ran.

Change-Id: I671bb43eddf41c7403de53bb4a223762101adc3c