Don't reset the local repo
Instead of "resetting" the local repo (git remote update,
git checkout master, git reset --hard, git clean -xfdq) before
merging each change, just fetch the remote ref for the branch
and check that out (as a detached head).  Or, if we are merging
a change that depends on another change in the queue, just check
that change out.

Change-Id: I0a9b839a0c75c04eca7393d7bb58cf89448b6494