Update patch set 1
Patch Set 1:

They didn't really address the current slowness we see in production; I constructed unrealistic tests based on an incomplete understanding of all of the factors, particularly around how the zuul repositories behave over a long period of time.  In reality, a fetch of a remote branch takes about the same time as a remote update, assuming that all branches are being kept relatively up-to-date, which they are in this case.  Moreover, by not tracking the remote branches (which is what the reverted optimizations do), git loses the ability to keep useful objects around.  Over time as the two repos drift, git may need to download more and more data for each fetch, which is what makes them 'bogus'.

Patch-set: 1
Label: Code-Review=0