Update patch set 3
Patch Set 3: (7 inline comments)

Looks good to me but I'm not leaving a vote because I'm not entirely sure I understand it all.

Does this open a possibility where a dependent change can pass when a change ahead of it fails? If so, would this end up with some tricky rebasing for the author?

Patch-set: 3
Label: Code-Review=0