Update patch set 3
Patch Set 3: I would prefer that you didn't merge this

(2 inline comments)

Soft -1, agreeing with Joshua's cleanup suggestion on gearman.py.

As for merging changes marked as dependent on other changes in Gerrit which didn't pass/merge, this still shouldn't be possible for two reasons... first, because the checkForChangesNeededBy() method is still being used to decide if it's whole; second, because currently we still rely on Gerrit to perform the eventual merge and it should flat out refuse when asked to do so (though relying on that alone wouldn't be great, of course).

Patch-set: 3
Label: Code-Review=-1
Label: Workflow=0