Update patch set 7
Patch Set 7:

Thanks for answering my questions. I think that makes sense now :-).

I actually liked the tests you had before where B -> A. Primarily because it checks that A still merges even though B (below it) fails.

Do we have a test that will ensure they are merged in the correct order?

E, D -> C -> B -> A, G -> F

With B failing (as before), will A still get merged and will it be ahead or behind F+G? Or, do we even care if A gets in before F? Is there a scenario where it gets merged F, A, G and if so will the dependents B and C care?

I suspect the above scenario is moot as I believe we can merge in that order and if it causes conflicts/failing tests then that is just upstream changing and the need for changes to be rebased but I wouldn't mind hearing your thoughts on this?

(/me hopes that makes sense, please ignore if not)

Patch-set: 7