Update patch set 7
Patch Set 7:

Whee that's a complicated comment!

So I changed 'B -> A' to 'B, A' because A is basically just there to ensure that whatever is happening is not happening at the head of the queue (otherwise the first of our dependent changes would try to merge).  We hold A's jobs until the end to make sure that the dependent changes we're looking at stay where they are in the queue.

I don't recall that we have a specific test to determine that if, in the case of B->A where B fails that A is merged.  I think that's because the scheduler algorithm has never looked behind to determine what to do, so it's hard to imagine such a case.  That doesn't mean it shouldn't be a test, of course.  But I'm not sure that requirement should be attached to this change.

To answer your questions about E, D -> C -> B -> A, G -> F:

A will be merged behind G and F.  F followed by A and G won't happen because the _order_ should never change, the 'moving' that happens is effectively only to omit some changes.

Patch-set: 7