Update patch set 2
Patch Set 2: Looks good to me (core reviewer)

This is really only used in two contexts: log messages and the status JSON.

The goal of this patch is to cause the status JSON to indicate that if a change has no discernable ID, then to have it be a null javascript value instead of the string 'None', which is weird.  In my mind that is a closer semantic match to what's going on (it is either a uniquely-identifying string value or null).

This change accomplishes that, but also changes what's passed to the log entries as well.  Instead of passing a string in all cases, they will be passed a reference to None in some cases.  Semantically, this also seems fairly similar, though I do easily see the appeal of a string-only api here.

The None reference serializes to the string 'None', so the log entries themselves should not change.  Given that, this seems like a fairly harmless way of implementing this; other approaches may be needlessly complex.  I don't feel strongly about this and am open to other suggestions, but lacking those, I'm okay with this.

Thanks, Alex!

Patch-set: 2
Label: Code-Review=+2