Clone repo if it is not on disk
* zuul/merger.py: Previously zuul's merger only checked an initialized
flag to determine if a repo should be cloned. If the state of the disk
changes under zuul the repo may need to be recloned even if the
initialized flag is set. Check if the repo exists on disk in addition to
the initialized flag to determine if the repo should be cloned to deal
with external state changes.

Change-Id: I14fb34a2916002cefef73e41ec9182a073d59ef3