Clone repo if it is not on disk
* zuul/merger.py: Previously zuul's merger only checked an initialized
flag to determine if a repo should be cloned. If the state of the disk
changes under zuul the repo may need to be recloned even if the
initialized flag is set. Check if the repo exists on disk to determine
if the repo should be cloned to deal with external state changes.

* tests/test_scheduler.py: Add a test that checks a repo is recloned
when removed from the filesystem under zuul.

* tests/fixtures/layout-repo-deleted.yaml: Layout fixture for new test
added above. This creates a new repo unused by any other tests so that
failures of this test do not interfere with other tests, they will run
independent of each other.

Change-Id: I14fb34a2916002cefef73e41ec9182a073d59ef3