Add the possibility to pass the zuul url
This patch adds the possibility to pass ZUUL_URL to jenkins, if
zuul_url is set on zuul conf file. Otherwise, it skip it, and
devstack-vm-gate-wrap.sh will use the default
http://zuul.openstack.org/p url.
This will enable other infrastructures to test changes with
their own zuul servers.
Documentation added in launchers.rst explaining this new option
in zuul.conf

Change-Id: I840423cc06fdfdacd301d30be3e0b3e589e563e9