Update patch set 3
Patch Set 3: Looks good to me (core reviewer)

The main concern over testing is to help you and others who might come to rely on this envvar. Since right now it defaults to our URL anyway, we don't set an override for it in configuration and won't notice if we accidentally break it. Tests for all the envvars would be nice, I think (so we can also avoid breaking ourselves), but that might be fodder for a separate patch anyway.

Patch-set: 3
Label: Code-Review=+2
Label: Workflow=0